Remote Control control code
The Eye's movement was controlled using geared DC brushless motors, and RC receiever, an H-bridge emulating motor controller, and a model car handhelp remote control transmitter. This code interprets the signal from the RC transmitter and turns into commands to be sent to the motor controller, thus controlling the movement of the eye.